chore: Type check refactoring
