+++ src/api/helper-types.ts
@@ -5,3 +5,4 @@ export type Nullable<T> = StrictNullable<T> | undefined;
 export type IsInstalled = (id: string) => boolean;
 
 export type TypeCheckFunction<T> = (value: unknown) => value is T;
+export type TypeCastCheckFunction<T> = <X = T>(value: unknown) => value is X;
