+++ src/api/options-manager.ts
@@ -4,7 +4,7 @@ import slash from 'slash';
 import { BundlibOptions } from './bundlib-options';
 import { OPTION_FILE_PATHS, PRODUCT_NAME } from './consts';
 import { error } from './errors';
-import { isString } from './type-check';
+import { isString } from './type-check/type-check';
 
 interface LoadedOptions {
   config: BundlibOptions;
