+++ src/api/validate/option-globals.ts
@@ -1,7 +1,7 @@
 import { GlobalsOptions } from '../bundlib-options';
 import { Nullable } from '../helper-types';
 import { keys, setProp } from '../helpers';
-import { isArray, isNull, isObject, isString } from '../type-check';
+import { isArray, isNull, isObject, isString } from '../type-check/type-check';
 
 export function isValidGlobals(value: unknown): value is GlobalsOptions {
   return isNull(value) || (
