+++ src/api/validate/validate-keys.ts
@@ -1,6 +1,5 @@
-import { Dictionary, StrictNullable } from '../helper-types';
+import { Dictionary, StrictNullable, TypeCheckFunction } from '../helper-types';
 import { keys } from '../helpers';
-import { InList } from './in-list';
 
 export function invalidKeys(object: Dictionary<unknown>, list: string[]): StrictNullable<string[]> {
   const invalid = keys(object).filter(
@@ -12,7 +11,7 @@ export function invalidKeys(object: Dictionary<unknown>, list: string[]): Strict
 export function keysInList<M extends string>(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   obj: Dictionary<any>,
-  inList: InList<M>,
+  inList: TypeCheckFunction<M>,
 ): obj is Partial<Record<M, unknown>> {
   return keys(obj).every(inList);
 }
